imgui: bind internal popup / tooltip / menu *Ex entry points (family)
Next imgui_internal.h family after #170 (rank 3 from the survey:
bindable-as-is, low cost). Bound via bind_imgui.das internal_allow_func
(regen only): OpenPopupEx, ClosePopupToLevel, ClosePopupsExceptModals,
IsPopupOpen (raw-ImGuiID overload), BeginPopupEx, BeginTooltipEx,
BeginTooltipHidden, BeginMenuEx, MenuItemEx. New internal enum
ImGuiTooltipFlags_ (one flag, OverridePrevious).

v2 surface: the four Begin*Ex pair with the public End*, so they are the
unbalance-proof with_popup_ex / with_tooltip_ex / with_tooltip_hidden /
with_menu_ex guards (imgui_scope_builtin.das, which is neither lint- nor
doc-gated, like with_focus_scope). The five one-shots are raw on
ALLOWED_IMGUI alongside the Render* batch. MenuItemEx is the
icon/shortcut escape hatch next to the wrapped menu_item widget.

Regenerated via the junction trick: net +9 makeExterns, chunk-reflow
verified (overflow into a new func_32 chunk, 0 functions lost),
CMakeLists registers func_32.cpp. EOL-only .inc churn reverted. Rebuilt
dasModuleImgui.shared_module (MSVC). New example
examples/features/internal_popup_menu_ex.das exercises all 9 (auto-opens
the raw-id popup so the popup path + MenuItemEx + IsPopupOpen + the menu
bar + with_tooltip_hidden all run headlessly; exit 0). Docs gate checked
locally (no Uncategorized); format --verify + lint clean on all changed
.das.
