+++ src/dasIMGUI.enum.decl.cast.inc
@@ -39,4 +39,5 @@ DAS_BIND_ENUM_CAST(ImDrawListFlags_);
 DAS_BIND_ENUM_CAST(ImFontAtlasFlags_);
 DAS_BIND_ENUM_CAST(ImGuiViewportFlags_);
 DAS_BIND_ENUM_CAST(ImGuiItemFlags_);
+DAS_BIND_ENUM_CAST(ImGuiTooltipFlags_);
 DAS_BIND_ENUM_CAST(ImGuiNavHighlightFlags_);
