+++ src/dasIMGUI.func.decl.inc
@@ -32,3 +32,4 @@ void initFunctions_28();
 void initFunctions_29();
 void initFunctions_30();
 void initFunctions_31();
+void initFunctions_32();
