+++ src/dasIMGUI.func.reg.inc
@@ -32,3 +32,4 @@ initFunctions_28();
 initFunctions_29();
 initFunctions_30();
 initFunctions_31();
+initFunctions_32();
